Fixed requirements, moved jet class options in experiment.py to a different class and file
