+++ requirements.txt
@@ -32,18 +32,8 @@ xformers==0.0.24
 torchdiffeq # for CFM sampling
 
 #weaver packages
-numpy>=1.19.5
-scipy>=1.5.2
-pandas>=1.1.3
-scikit-learn>=1.0.1
-matplotlib>=3.3.0
 tqdm>=4.41.0
 PyYAML>=5.4
 awkward0>=0.15.5
 uproot>=4.2.0,<5.2.0
 awkward>=1.8.0
-vector>=0.9.0
-lz4>=3.1.0
-xxhash>=1.4.4
-tables>=3.6.1
-tensorboard>=2.2.0
